3.20.00

+++ home/dicom/age.bash
@@ -1,7 +1,7 @@
 #!/bin/bash
-# Version 3.00.00b14
-# Build 2
-# 2015-11-23
+# Version 3.02.02
+# Build 3
+# 2018-12-26
 # License GPLv3
 
 RECEIVERS=`cat /etc/primal/primal.conf|grep "<scp"|cut -c5-|cut -d ">" -f1|tr "\n" " "`
@@ -17,7 +17,12 @@ do
 
 	DIRFOUND=`find $PRISENT -name "$1_*" -type d -ctime +$PRIRET|wc -l`
 	echo "Receiver: $1   Number of directories found that are to be removed:  $DIRFOUND   Days of retentnion:  $PRIRET"
-	find $PRISENT -name "$1_*" -type d -ctime +$PRIRET -exec rm -fr {} \;
+	if [ "$PRIARCHTYPE0" == "S3" ] || [ "$PRIARCHTYPE0" == "disk" ]
+	then
+		find $PRISENT -name "$1_*" -type d -ctime +$PRIRET -exec /home/dicom/age2.bash {} \;
+	else
+		find $PRISENT -name "$1_*" -type d -ctime +$PRIRET -exec rm -fr {} \;
+	fi
 	DIRFOUND=`find $PRISENT -name "$1_*" -type d -ctime +$PRIRET|wc -l`
 	echo "Receiver: $1   Number of directories found after processing (should be zero):  $DIRFOUND   Days of retentnion:  $PRIRET"
 
@@ -41,7 +46,7 @@ do
 		DIRNAME=`ls -1str $PRISENT|head -3|tail -1|tr -s " "|cut -d " " -f2`
 		if [ "$DIRNAME" != "" ] && [ "$DIRNAME" != " " ]
 		then
-			echo "update image set ilocation = NULL where puid = '$DIRNAME';"|mysql -u root primal
+			echo "update image set ilocation = NULL where puid = '$DIRNAME';"|$DBCONN
 			rm -fr $PRISENT/$DIRNAME
 		fi
 		FILLLEVEL=`df -h $PRISENT|tail -1|tr -s " "|cut -d " " -f5|cut -d "%" -f1`
